+++ muttley/src/main/java/com/project/muttley/domain/event/dto/EventRequestDTO.java
@@ -15,7 +15,7 @@ public record EventRequestDTO(
     @NotNull(message = "Data inicial é obrigatória") LocalDate startDate,
     LocalDate endDate,
     Integer workload,
-    Long points,
+    Integer points,
     @NotBlank(message = "Tipo do evento é obrigatório") String typeId,
     String subject,
     @MaxKeywords(max = 10) String keywords,
